Save and Load

+++ NotJetpackJoyrideProject/Assets/Scripts/DataPersistence/Data/Data.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class GameData
 {
-    public float HighestScore;
+    public int HighestScore;
 
     public GameData()
     {
